server/requirements.txt
absl-py==0.1.11
argh==0.26.2
astor==0.6.2
auxlib==0.0.42
beautifulsoup4==4.6.0
bleach==1.5.0
click==6.7
colorama==0.3.9
conda==4.2.7
cssmin==0.2.0
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12.2
Flask-Assets==0.12
Flask-Cors==3.0.3
Flask-Psycopg2==1.3
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
pathtools==0.1.2
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pygments==2.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.5
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tifffile==0.14.0
tornado==5.0.1
tqdm==4.19.8
traitlets==4.3.2
watchdog==0.8.3
wcwidth==0.1.7
webassets==0.12.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
win-unicode-console==0.5
WTForms==2.1
